Remove InterpretercoreInferShapeContext
